+++ CHANGELOG.md
@@ -12,6 +12,10 @@ before starting to add changes. Use example [placed in the end of the page](#exa
 ## [Unreleased]
 
 - Avoid double-saving submissions when handling name and address protection.
+- [PR-191](https://github.com/OS2Forms/os2forms/pull/191)
+  Re-throws exception to ensure failed status during Maestro notification job.
+- [PR-202](https://github.com/OS2Forms/os2forms/pull/202)
+  - Removed non-digits from recipient id in Maestro digital post notifications.
 - [PR-189](https://github.com/OS2Forms/os2forms/pull/189)
   - Added support for MeMo 1.2 and added additional validation of MeMo actions.